pyproject.toml
[tool.poetry]
name = "cloudflare-sync-action"
version = "0.0.0"
description = "A GitHub action that sync the CNAME DNS records in Cloudflare with local files."
authors = ["George Cheng <gerhut@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
cloudflare = "^2.8.13"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
python-dotenv = "^0.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
